Organize sample components
